pyproject.toml
[project]
name = "smart-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agentops>=0.4.16",
    "aiofiles>=24.1.0",
    "openai-agents>=0.0.19",
    "tavily-python>=0.7.8",
]
